# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Firat Api",
        "version": "0.1"
    },
    "paths": {
        "/api/v1/announcement/all": {
            "get": {
                "tags": [
                    "Duyuru"
                ],
                "description": "Duyuruları Getir",
                "operationId": "8396949637b91c4430172b8fe2a7702d",
                "responses": {
                    "200": {
                        "description": "İşlem başarılı"
                    },
                    "400": {
                        "description": "Geçersiz istek"
                    },
                    "401": {
                        "description": "Yetkisiz"
                    },
                    "500": {
                        "description": "Sunucu hatası"
                    }
                }
            }
        },
        "/api/v1/announcement/get": {
            "get": {
                "tags": [
                    "Duyuru"
                ],
                "description": "Duyuruları id'ye göre getir",
                "operationId": "7115d5bf8fdfa50c7a3c8aeddc81d117",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Parametre açıklaması",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/v1/announcement/create": {
            "post": {
                "tags": [
                    "Duyuru"
                ],
                "description": "Duyuru Oluştur",
                "operationId": "e14b779997a9fd41a47f0825134dac67",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "İşlem başarılı"
                    },
                    "400": {
                        "description": "Geçersiz istek"
                    },
                    "401": {
                        "description": "Yetkisiz"
                    },
                    "500": {
                        "description": "Sunucu hatası"
                    }
                }
            }
        },
        "/api/v1/announcement/update": {
            "post": {
                "tags": [
                    "Duyuru"
                ],
                "description": "Duyuru Güncelle",
                "operationId": "700110217a525de6883622e8d0dd8935",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "İşlem başarılı"
                    },
                    "400": {
                        "description": "Geçersiz istek"
                    },
                    "401": {
                        "description": "Yetkisiz"
                    },
                    "500": {
                        "description": "Sunucu hatası"
                    }
                }
            }
        },
        "/api/v1/announcement/delete": {
            "post": {
                "tags": [
                    "Duyuru"
                ],
                "description": "Duyuru Sil",
                "operationId": "0b5912d4d5caade371af46b43fbb2cf4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "İşlem başarılı"
                    },
                    "400": {
                        "description": "Geçersiz istek"
                    },
                    "401": {
                        "description": "Yetkisiz"
                    },
                    "500": {
                        "description": "Sunucu hatası"
                    }
                }
            }
        },
...
    }
}